Refactor extension.ts
